+++ khal/controllers.py
@@ -27,12 +27,12 @@ import re
 import textwrap
 from collections import OrderedDict, defaultdict
 from shutil import get_terminal_size
-from typing import List, Optional
+from typing import Callable, List, Optional
 
 import pytz
 from click import confirm, echo, prompt, style
 
-from khal import __productname__, __version__, calendar_display, parse_datetime, utils
+from khal import __productname__, __version__, calendar_display, parse_datetime
 from khal.custom_types import (
     EventCreationTypes,
     LocaleConfiguration,
@@ -49,6 +49,7 @@ from .icalendar import cal_from_ics, split_ics
 from .icalendar import sort_key as sort_vevent_key
 from .khalendar.vdir import Item
 from .terminal import merge_columns
+from .utils import human_formatter, json_formatter
 
 logger = logging.getLogger('khal')
 
@@ -95,6 +96,7 @@ def calendar(
     bold_for_light_color: bool=True,
     env=None,
     ):
+
     term_width, _ = get_terminal_size()
     lwidth = 27 if conf['locale']['weeknumbers'] == 'right' else 25
     rwidth = term_width - lwidth - 4
@@ -170,24 +172,27 @@ def get_events_between(
     locale: dict,
     start: dt.datetime,
     end: dt.datetime,
-    agenda_format: str,
+    formatter: Callable,
     notstarted: bool,
     env: dict,
-    width: Optional[int],
-    seen,
     original_start: dt.datetime,
+    seen=None,
+    colors: bool = True,
 ) -> List[str]:
     """returns a list of events scheduled between start and end. Start and end
     are strings or datetimes (of some kind).
 
     :param collection:
+    :param locale:
     :param start: the start datetime
     :param end: the end datetime
-    :param agenda_format: a format string that can be used in python string formatting
-    :param env: a collection of "static" values like calendar names and color
+    :param formatter: the formatter (see :class:`.utils.human_formatter`)
     :param nostarted: True if each event should start after start (instead of
-    be active between start and end)
+      be active between start and end)
+    :param env: a collection of "static" values like calendar names and color
     :param original_start: start datetime to compare against of notstarted is set
+    :param seen:
+    :param colors:
     :returns: a list to be printed as the agenda for the given days
     """
     assert not (notstarted and not original_start)
@@ -218,23 +223,20 @@ def get_events_between(
             continue
 
         try:
-            event_string = event.format(agenda_format, relative_to=(start, end), env=env)
+            event_attributes = event.attributes(relative_to=(start, end), env=env, colors=colors)
         except KeyError as error:
             raise FatalError(error)
 
-        if width:
-            event_list += utils.color_wrap(event_string, width)
-        else:
-            event_list.append(event_string)
+        event_list.append(event_attributes)
         if seen is not None:
             seen.add(event.uid)
 
-    return event_list
+    return formatter(event_list)
 
 
 def khal_list(
     collection,
-    daterange: Optional[List[str]]=None,
+    daterange: Optional[List[str]] = None,
     conf: Optional[dict] = None,
     agenda_format=None,
     day_format: Optional[str]=None,
@@ -243,6 +245,7 @@ def khal_list(
     width: Optional[int] = None,
     env=None,
     datepoint=None,
+    json: Optional[List] = None,
 ):
     """returns a list of all events in `daterange`"""
     assert daterange is not None or datepoint is not None
@@ -251,6 +254,14 @@ def khal_list(
     # because empty strings are also Falsish
     if agenda_format is None:
         agenda_format = conf['view']['agenda_event_format']
+
+    json_mode = json is not None and len(json) > 0
+    if json_mode:
+        formatter = json_formatter(json)
+        colors = False
+    else:
+        formatter = human_formatter(agenda_format, width)
+        colors = True
 
     if daterange is not None:
         if day_format is None:
@@ -296,13 +307,13 @@ def khal_list(
         else:
             day_end = dt.datetime.combine(start.date(), dt.time.max)
         current_events = get_events_between(
-            collection, locale=conf['locale'], agenda_format=agenda_format, start=start,
+            collection, locale=conf['locale'], formatter=formatter, start=start,
             end=day_end, notstarted=notstarted, original_start=original_start,
             env=env,
             seen=once,
-            width=width,
+            colors=colors,
         )
-        if day_format and (conf['default']['show_all_days'] or current_events):
+        if day_format and (conf['default']['show_all_days'] or current_events) and not json_mode:
             if len(event_column) != 0 and conf['view']['blank_line_before_day']:
                 event_column.append('')
             event_column.append(format_day(start.date(), day_format, conf['locale']))
@@ -314,7 +325,7 @@ def khal_list(
 
 def new_interactive(collection, calendar_name, conf, info, location=None,
                     categories=None, repeat=None, until=None, alarms=None,
-                    format=None, env=None, url=None):
+                    format=None, json=None, env=None, url=None):
     info: EventCreationTypes
     try:
         info = parse_datetime.eventinfofstr(
@@ -381,6 +392,7 @@ def new_interactive(collection, calendar_name, conf, info, location=None,
         format=format,
         env=env,
         calendar_name=calendar_name,
+        json=json,
     )
     echo("event saved")
 
@@ -390,7 +402,7 @@ def new_interactive(collection, calendar_name, conf, info, location=None,
 
 def new_from_string(collection, calendar_name, conf, info, location=None,
                     categories=None, repeat=None, until=None, alarms=None,
-                    url=None, format=None, env=None):
+                    url=None, format=None, json=None, env=None):
     """construct a new event from a string and add it"""
     info = parse_datetime.eventinfofstr(
         info, conf['locale'],
@@ -406,7 +418,15 @@ def new_from_string(collection, calendar_name, conf, info, location=None,
         'alarms': alarms,
         'url': url,
     })
-    new_from_dict(info, collection, conf=conf, format=format, env=env, calendar_name=calendar_name)
+    new_from_dict(
+        info,
+        collection,
+        conf=conf,
+        format=format,
+        env=env,
+        calendar_name=calendar_name,
+        json=json,
+    )
 
 
 def new_from_dict(
@@ -416,6 +436,7 @@ def new_from_dict(
     calendar_name: Optional[str]=None,
     format=None,
     env=None,
+    json=None,
 ) -> Event:
     """Create a new event from arguments and save in vdirs
 
@@ -437,9 +458,13 @@ def new_from_dict(
         )
 
     if conf['default']['print_new'] == 'event':
-        if format is None:
-            format = conf['view']['event_format']
-        echo(event.format(format, dt.datetime.now(), env=env))
+        if json is None or len(json) == 0:
+            if format is None:
+                format = conf['view']['event_format']
+            formatter = human_formatter(format)
+        else:
+            formatter = json_formatter(json)
+        echo(formatter(event.attributes(dt.datetime.now(), env=env)))
     elif conf['default']['print_new'] == 'path':
         assert event.href
         path = os.path.join(collection._calendars[event.calendar]['path'], event.href)
@@ -504,7 +529,7 @@ def edit_event(event, collection, locale, allow_quit=False, width=80):
                 collection.delete(event.href, event.etag, event.calendar)
                 return True
         elif choice == "datetime range":
-            current = event.format("{start} {end}", relative_to=now)
+            current = human_formatter("{start} {end}")(event.attributes(relative_to=now))
             value = prompt("datetime range", default=current)
             try:
                 start, end, allday = parse_datetime.guessrangefstr(ansi.sub('', value), locale)
@@ -588,7 +613,8 @@ def edit(collection, search_string, locale, format=None, allow_past=False, conf=
                 continue
             elif not event.allday and event.end_local < now:
                 continue
-        event_text = textwrap.wrap(event.format(format, relative_to=now), term_width)
+        event_text = textwrap.wrap(human_formatter(format)(
+            event.attributes(relative_to=now)), term_width)
         echo(''.join(event_text))
         if not edit_event(event, collection, locale, allow_quit=True, width=term_width):
             return
@@ -640,7 +666,7 @@ def import_event(vevent, collection, locale, batch, format=None, env=None):
             if item.name == 'VEVENT':
                 event = Event.fromVEvents(
                     [item], calendar=collection.default_calendar_name, locale=locale)
-                echo(event.format(format, dt.datetime.now(), env=env))
+                echo(human_formatter(format)(event.attributes(dt.datetime.now(), env=env)))
 
     # get the calendar to insert into
     if not collection.writable_names:
@@ -697,4 +723,4 @@ def print_ics(conf, name, ics, format):
     echo(f'{len(vevents)} events found in {name}')
     for sub_event in vevents:
         event = Event.fromVEvents(sub_event, locale=conf['locale'])
-        echo(event.format(format, dt.datetime.now()))
+        echo(human_formatter(format)(event.attributes(dt.datetime.now())))