+++ tests/utils_test.py
@@ -1,6 +1,7 @@
 """testing functions from the khal.utils"""
 import datetime as dt
 
+from click import style
 from freezegun import freeze_time
 
 from khal import utils
@@ -119,3 +120,9 @@ def test_get_weekday_occurrence():
     assert utils.get_weekday_occurrence(dt.date(2017, 5, 8)) == (0, 2)
     assert utils.get_weekday_occurrence(dt.date(2017, 5, 28)) == (6, 4)
     assert utils.get_weekday_occurrence(dt.date(2017, 5, 29)) == (0, 5)
+
+
+def test_human_formatter_width():
+    formatter = utils.human_formatter('{red}{title}', width=10)
+    output = formatter({'title': 'morethan10characters', 'red': style('', reset=False, fg='red')})
+    assert output.startswith('\x1b[31mmoret\x1b[0m')